refactor(chat): déplacer FavoriteConversationResult dans dto/ et corriger imports
- Crée application/dto/favorite_conversation_dto.py avec FavoriteConversationResult en @dataclass
- Supprime la définition inline dans conversation_repository.py (déplacée dans dto/)
- Déplace l'import tardif de ConversationNotFoundError en haut de sqlalchemy_conversation_repository.py
- InMemoryConversationRepository accepte un dict optionnel project_names pour les tests
