+++ server/src/raggae/application/dto/favorite_conversation_dto.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+from raggae.domain.entities.conversation import Conversation
+
+
+@dataclass
+class FavoriteConversationResult:
+    """Résultat enrichi d'une conversation favorite avec le nom du projet associé."""
+
+    conversation: Conversation
+    project_name: str
